requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.0.1
opencv-python==4.5.5.64
numpy==1.21.5